chore : 폴더 변경 및 주석 수정
