+++ src/Pages/ProjectPage/ProjectPage.tsx
@@ -3,14 +3,14 @@ import styled from 'styled-components';
 import { createGlobalStyle } from 'styled-components';
 import ProjectInfo from './ProjectInfo';
 import ProfileList from './ProfileList/ProfileList';
-import MySchedule from './Schedule/MySchedule';
-import TeamSchedule from './Schedule/TeamSchedule';
+import MySchedule from './Schedule/my/MySchedule';
+import TeamSchedule from './Schedule/team/TeamSchedule';
 import Alarm from './Alarm';
-import MemberSchedule from './Schedule/MemberSchedule';
+import MemberSchedule from './Schedule/member/MemberSchedule';
 import ArrowDropUpIcon from '@mui/icons-material/ArrowDropUp';
 import ArrowDropDownIcon from '@mui/icons-material/ArrowDropDown';
 import { useState } from 'react';
-import MemberScheduleGrid from './Schedule/MemberScheduleGrid';
+import MemberScheduleGrid from './Schedule/member/MemberScheduleGrid';
 import Info from '../ProjectMakePage/Info';
 import Project from '../MainPage/components/Project';
 import ProjectCalendar from '../ProjectMakePage/projectcalendar';
