+++ src/hooks/apis/queries/schedule/useScheduleMemberQuery.ts
@@ -2,9 +2,9 @@ import { useQuery } from '@tanstack/react-query';
 import { QUERY_KEY } from '#/constants/queryKey';
 import { getScheduleMember } from '#/apis/schedule';
 /**
- * GET /v1/schedules/all
+ * GET /v1/projects/{projectId}/members/{memberId}/schedules
  *
- * 시간표 전체를 조회하는 api 입니다.
+ * 프로젝트 안에서 특정 유저가 작성한 시간표 출력하는 api 입니다.
  */
 const useScheduleMemberQuery = (projectId: number, memberId: number) => {
     return useQuery({
